♻️refactor : createdAt 타입을 Int? -> Double? 로 변경

+++ Diary/DiaryData.swift
@@ -10,7 +10,7 @@ import Foundation
 struct DiaryData: Decodable {
     let title: String?
     let body: String?
-    let createdAt: Int?
+    let createdAt: Double?
     
     private enum CodingKeys: String, CodingKey {
         case title, body
